Fixing schema type

+++ backend/back.js
@@ -6,7 +6,7 @@ const { MongoClient } = require('mongodb');
 const pkg = require('csvtojson');
 const bodyParser = require('body-parser');
 const {csv} = pkg;
-
+let count = 1;
 
 app.use(cors()); 
 app.use(bodyParser.json());
@@ -42,21 +42,24 @@ app.post('/upload', (req, res)=>{
 app.post('/schema', (req, res)=>{
     // console.log("OKAY RECIEVED")
     const data = req.body;
-    console.log(data[0])
-    const result = data[0].reduce((acc, obj) => {
-      Object.entries(obj).forEach(([key, value]) => {
-        if (!acc[key]) acc[key] = [];
-        acc[key].push(value);
-      });
-      return acc;
-    }, {});
-    
-    buildCollection(getClient(), result).catch(console.dir).then(()=>{
-      runConstraints(getClient(), data[0]).catch(console.dir).then(()=>{
-        res.json({"message":"gg"})
+    data.forEach((dataItem) =>{
+      const result = dataItem.reduce((acc, obj) => {
+        Object.entries(obj).forEach(([key, value]) => {
+          if (!acc[key]) acc[key] = [];
+          acc[key].push(value);
+        });
+        return acc;
+      }, {});
+      console.log(result)
+      count+=1
+      buildCollection(getClient(), result, count).catch(console.dir).then(()=>{
+        runConstraints(getClient(), dataItem).catch(console.dir).then(()=>{
+          
+        })
       })
-    })
-    // console.log(req.body);
+      // console.log(result);
+    });
+    res.json({"message":"gg"})
 })
 
 
@@ -77,7 +80,6 @@ async function queryValues(client){
       const table = database.collection("temporary");
       const obj = await table.findOne();
       const keys = Object.keys(obj);
-      console.log(keys)
       return keys
   } finally{
       await client.close();
@@ -97,16 +99,84 @@ async function insertValues(client, documents) {
   }  
 }
 
-async function buildCollection(client, columns) {
+async function buildCollection(client, columns, count) {
   try {
       const database = client.db("testDB");
       const haiku = database.collection("temporary");
-      const name = database.collection("schema1");
+      const collectionName = "schema"+count;
+      const name = database.collection(collectionName);
+      // await name.insertOne({});
       const projection = columns.columnName.reduce((acc, curr) => ({ ...acc, [curr]: 1 }), {});
+
+      // This is to validate the structure of the collection
+
+      // const property = {};
+      // // const typ = columns.dataType.map(DataType => ({ bsonType: DataType }));
+      // const typ = {bsonType: 'string'}
+      // // Loop through each column name and add its data type to the properties object
+      // console.log(typ);
+      // columns.columnName.forEach((columnName, index) => {
+      //   property[columnName] = typ;
+      // });
+      // console.log(property)
+      // const s = {
+      //   $jsonSchema: {
+      //   bsonType: "object",
+      //   required: columns.columnName,
+      //   properties: property,
+      //   }
+      // };
+      // console.log(s)
+      // const val = {validator:s}
+
+      // const name = await database.createCollection(collectionName, val);
+
       const data = await haiku.find({}, {projection}).toArray()
       await name.insertMany(data, (err) =>{
           if (err) throw err;
       })
+      console.log("HOYA")
+      for(let i=0; i<columns.columnName.length; i++){
+        console.log("HOYA")
+        console.log(columns.columnName[i])
+        console.log(columns.tempColName[i])
+        const res = await name.updateMany(
+          {}, {$rename: {[columns.columnName[i]]:columns.tempColName[i]}}
+        )
+        console.log(res)
+      }
+
+      //To typecase (Hardcoded for input collection of type)
+      // {
+      //   columnName: [ 'age' ],
+      //   dataType: [ 'Integer' ],
+      //   pk: [ 'f' ],
+      //   nc: [ 'f' ],
+      //   uc: [ 'f' ],
+      //   fk: [ 'f' ],
+      //   tempColName: [ 'ahr' ]
+      // }
+      await name.updateMany({}, { $set: { "ahr": { $toInt: "$ahr" } } });
+
+      // for(let i=0; i<columns.columnName.length; i++){
+      //   const attributeName = columns.columnName[i];
+      //   console.log(`$${attributeName}`)
+      //   switch (columns.dataType[i]) {
+      //     case 'Integer':
+      //       console.log("YEAHHH");
+      //       const res = await name.updateMany(
+      //         { [attributeName]: { $not: { $type: "number" } } },
+      //         { $set: { [attributeName]: { $toInt: `$age` } } }
+      //       )
+      //       console.log(res)
+      //       break;
+      //     // Add more cases for other data types as needed
+      //     default:
+      //       break;
+      //   }
+      // }
+
+
   } finally {
       await client.close();
   }
@@ -115,17 +185,45 @@ async function buildCollection(client, columns) {
 async function runConstraints(client, columns) {
   try {
     const database = client.db("testDB");
-    const name = database.collection("schema1");
+    const collectionName = "schema"+count;
+    const name = database.collection(collectionName);
 
     for (let i = 0; i < columns.length; i++) {
       const item = columns[i];
 
-      if (item.nc == 1) {
+      if (item.nc == 1 && item.replaceWith == "none") {
         try {
           const query = { [item.columnName]: { $type: "null" } };
           const result = await name.deleteMany(query);
-          console.log(result);
+    
         } catch (err) {
+          console.error(err);
+        }
+      }
+      if(item.nc == 1 && item.replaceWith == "mean"){
+        try{
+          const avgResult = await name.aggregate([
+            {
+              $group: {
+                _id: null,
+                avgValue: { $avg: "$age" }
+              }
+            },
+            {
+              $project: {
+                _id: 0,
+                avgValue: 1
+              }
+            }
+          ]).toArray();
+
+          const result = await name.updateMany(
+            { age: null },
+            { $set: { age: avgResult[0].avgValue } }
+          );
+          console.log(`${result.modifiedCount} documents updated.`);
+
+        }catch(err){
           console.error(err);
         }
       }
